admin/requirements.txt
fastapi
uvicorn
requests
pyyaml
python-multipart
pydantic
pytest
